refactor: move tests back to original spot
